feat: add meta descriptions across pages
